fix: widen phisteps/mosaic_domains from short to int in CUDA kernel
goniometerParams.phisteps and crystalParams.mosaic_domains were declared
short, along with the phi/mosaic hot-loop counters (phi_tic, mos_tic).
The host driver and the CPU reference both carry these as int, so a value
above 32767 was narrowed at the struct assignment and wrapped to a <= 0
loop bound: the phi and mosaic loops ran zero iterations and the kernel
returned an empty image (max_I = 0) with no error or warning. This is
reachable in normal use -- phisteps is usually derived as ceil(osc/phistep),
e.g. -osc 360 -phistep 0.01 = 36000 steps.

Widen the two struct fields and the two loop counters to int, matching the
host and CPU reference across the full int domain. Register-neutral
(55->55 single, 64->64 df64, no spill) and byte-identical for in-range
values on both precisions.

+++ cuda/src/nanoBraggCUDA.cu
@@ -157,14 +157,14 @@ struct crystalParams {
     float default_F;
     float dmin;
     shapetype xtal_shape;
-    short mosaic_domains;
+    int mosaic_domains;
     structureFactorParams fhklParams;
 };
 
 struct goniometerParams {
     float phi0;
     float phistep;
-    short phisteps;
+    int phisteps;
     float spindle_vector[VECTOR_SIZE];
 };
 
@@ -985,10 +985,10 @@ static __global__ void nanoBraggSpotsCUDAKernel(const detectorParams * __restric
                         }
 
                         /* sweep over phi angles */
-                        for (short phi_tic = 0; phi_tic < goniometer->phisteps; ++phi_tic) {
+                        for (int phi_tic = 0; phi_tic < goniometer->phisteps; ++phi_tic) {
 
                             /* enumerate mosaic domains */
-                            for (short mos_tic = 0; mos_tic < crystal->mosaic_domains; ++mos_tic) {
+                            for (int mos_tic = 0; mos_tic < crystal->mosaic_domains; ++mos_tic) {
 
                                 /* Outputs of the lattice-shape block below, all single precision:
                                    the Miller indices h,k,l (the round crystal shapes need their
